feat(.circleci/config.yml): add sally job and bobby_sally workflow
This commit introduces a new job called "sally" and a workflow
named "bobby_sally" to the CircleCI configuration. The "sally" job
is configured to use the circleci/node@7.1.0 Docker image and
executes a simple "Double Trouble" echo command. The "bobby_sally"
workflow orchestrates the execution of both the "bobby" and "sally"
jobs.